ルックアップゲート攻略 ルックアップゲート攻略に使う閾値の取得をする処理をキャリブレーションに追加。UP_BUTTONで設定する

+++ app/src/ultrasonic.c
@@ -9,7 +9,7 @@
  * @return distance ルックアップゲートとの間合い
  */
 int16_t Ultrasonic_get_distance(){
-	int16_t distance = ev3_ultrasonic_sensor_get_distance(ULTRASONIC_SENSOR_P);
+	int16_t distance = (int)ev3_ultrasonic_sensor_get_distance(ULTRASONIC_SENSOR_P);
 	return distance;
 }
